+++ example/test/flutter_test_config.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:golden_test/cached_network_image.dart';
 import 'package:golden_test/golden_test.dart';
 import 'package:golden_test_example/l10n/app_localizations.dart';
 
@@ -30,6 +31,8 @@ Future<void> testExecutable(FutureOr<void> Function() testMain) async {
     GlobalWidgetsLocalizations.delegate,
     GlobalCupertinoLocalizations.delegate,
   ];
+
+  setupGoldenTestCachedNetworkImage();
 
   return testMain();
 }