+++ lib/golden_test.dart
@@ -3,5 +3,6 @@ library;
 export 'src/config.dart';
 export 'src/device.dart';
 export 'src/golden_test.dart';
+export 'src/network_image_stub.dart';
 export 'src/text_scale.dart';
 export 'src/tools.dart';